Implement minimize_regs as an analysis
