+++ calyx/src/passes/mod.rs
@@ -9,7 +9,6 @@ mod dead_cell_removal;
 mod externalize;
 mod go_insertion;
 mod inliner;
-mod live_range_analysis;
 mod minimize_regs;
 mod papercut;
 mod remove_external_memories;
@@ -26,7 +25,6 @@ pub use dead_cell_removal::DeadCellRemoval;
 pub use externalize::Externalize;
 pub use go_insertion::GoInsertion;
 pub use inliner::Inliner;
-pub use live_range_analysis::LiveRangeAnalysis;
 pub use minimize_regs::MinimizeRegs;
 pub use papercut::Papercut;
 pub use remove_external_memories::RemoveExternalMemories;
