+++ src/main.rs
@@ -15,7 +15,7 @@ use pass_manager::PassManager;
 use passes::{
     ClkInsertion, CollapseControl, CompileControl, CompileEmpty,
     ComponentInterface, DeadCellRemoval, Externalize, GoInsertion, Inliner,
-    LiveRangeAnalysis, MinimizeRegs, Papercut, RemoveExternalMemories,
+    MinimizeRegs, Papercut, RemoveExternalMemories,
     ResourceSharing, StaticTiming, WellFormed,
 };
 use std::io::stdin;
@@ -43,17 +43,10 @@ fn construct_pass_manager() -> FutilResult<PassManager> {
     register_pass!(pm, ClkInsertion);
     register_pass!(pm, ResourceSharing);
     register_pass!(pm, DeadCellRemoval);
+    register_pass!(pm, MinimizeRegs);
 
-    // custom register pass
-    let register_removal_pass_f: pass_manager::PassClosure = Box::new(|ctx| {
-        let analysis = LiveRangeAnalysis::do_pass_default(ctx)?;
-        MinimizeRegs::new(analysis).do_pass(ctx)?;
-        Ok(())
-    });
-    pm.add_pass(MinimizeRegs::name().to_string(), register_removal_pass_f)?;
 
     // Register aliases
-    // TODO: Add resource sharing.
     register_alias!(
         pm,
         "all",
